Split out AggregatedStats, ChangedEntityLocale, Comment & PermissionChangelog
